+++ CHANGELOG.md
@@ -58,6 +58,8 @@ These changes are available on the `master` branch, but have not yet been releas
     ([#2623](https://github.com/Pycord-Development/pycord/pull/2623))
 - Added new `Subscription` object and related methods/events.
   ([#2564](https://github.com/Pycord-Development/pycord/pull/2564))
+- Added ability to change the API's base URL with `Route.API_BASE_URL`.
+  ([#2714](https://github.com/Pycord-Development/pycord/pull/2714))
 
 ### Fixed